requirements.txt
Django==2.2
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.6.1
pycountry==18.12.8